requirements.txt
fastapi
uvicorn
sqlalchemy
aioredis
python-dotenv
alembic
pytest
psycopg2-binary  # Driver para PostgreSQL
